+++ CHANGELOG.md
@@ -2,6 +2,7 @@ Change format:
 - {type} {PR}: {description} (#{github-username})
 
 ## Pending release
+- Add #18: style attribute support (@es-sayers)
 - Add #17: accept-charset attribute support (@es-sayers)
 
 ## 1.0.0 - August 13, 2024